requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
discord-webhook==1.3.0
Django==4.2.4
django-cron==0.6.0
django-ipware==5.0.0
djangorestframework==3.14.0
idna==3.4
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
typing_extensions==4.7.1
urllib3==2.0.4
